Move LargeDataArray into Examples
